Notifications & permission requests implemented, added AppSettings class, added a probability indicator
Notifications & permission requests implemented. Notifier now observes MainActivity lifecycle - notifications are only sent when the app is in the background (or in a view other than the main view) .

Added AppSettings class for retrieving user settings.

Added a probability indicator to the main view, tied to the value of magnetic activity.