Added an update timer, implemented a min. interval for notifications, and a periodic Worker for notifications. Planned features are now completed.
Added a timer which executes periodic data fetching. The timer runs in WorkController and enqueues an UpdateWorker every 10 minutes. Time remaining to next update is shown to user in MainActivity. MainModel is subscribed to WorkController through the TimerSubscriber interface to receive remaining time on every tick, which is once per second. MainModel then updates its MutableLiveData which is reflected in MainActivity.

Implemented a minimum interval for notifications. UpdateWorker checks if a user-defined amount of time has passed since the last notification before sending a new one.

Added  PeriodicUpdateWorker for sending notifications when the app has been killed or not yet started after device reboot.

Most features are now completed and app is ready for 1.0 release. Finnish translation coming in 1.1.