Refactored work scheduling, improved Worker error handling, PeriodicUpdateWorker cache check
Refactored WorkController for more robust Worker scheduling. Added triggerUpdate() so that MainModel can request a data re-fetch if the app is in an error state when the user resumes using the app.

Added error handling for connection errors. This is especially helpful on devices which restrict background network traffic to optimize battery usage.

PeriodicUpdateWorker now checks for fresh data in the cache before attempting to fetch new data.